requirements.txt
keyboard
pyperclip
pyttsx3
pygame
pystray
Pillow

